requirements.txt
vanna[chromadb,openai]==0.7.4
chromadb==0.5.3
sentence-transformers==3.0.1
psycopg[binary]>=3.2.2
streamlit==1.38.0
python-dotenv==1.0.1
pandas==2.2.2
sqlalchemy==2.0.32
